// src/app/models/User.ts
export interface IUser {
    uid?: string;
    name?: string;
    email?: string;
    password?: string;
    userType?: string;
    avatarUrl?: string;
    mobile?: string;
    documentUrl?: string;
    mycoach?: string;
    languageId?: string;
    education?: string;
    businessinfo?: any;
    balanceSheet?: any;
    product?: any;
    loanApplication?: any;
}